tidy(restructuring): move fuzzy-matching helpers to src/matching.rs
